HR-Agent-MatchMaking/src/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pinecone-client==3.0.0
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.7.2
scrapy==2.11.0
selenium==4.15.0
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
scikit-learn==1.3.2
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.2.1
python-multipart==0.0.6
aiohttp==3.9.1
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
